edit and tour buttons setup
