+++ client/src/App.css
@@ -23,6 +23,7 @@
   display: flex;
   grid-column: width;
   grid-row: header;
+  z-index: 100;
 }
 
 .main {
