+++ client/src/components/profile-page/Tour.js
@@ -1,15 +1,14 @@
-import Navigation from "./Navigation";
+import Navigation from "../Navigation";
+import MainContent from "./MainContent";
 
 export default function Tour() {
   return (
     <div className="grid-layout-app">
       <header className="header">
-        <h1>
-          <span className="cover-title">tour</span>
-        </h1>
+        <h1 className="cover-title">tour</h1>
       </header>
       <main className="main tour-page">
-        <div className="tour-title">Welcome to the tour!</div>
+        <MainContent />
       </main>
       <footer className="footer">
         <Navigation />
